pyproject.toml
[project]
name = "helix-hands-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "cerebras-cloud-sdk>=1.50.1",
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "markdownify>=1.2.0",
    "markitdown[all]>=0.1.3",
    "mcp[cli]>=1.15.0",
    "openai>=1.109.1",
    "playwright>=1.55.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "youtube-transcript-api>=1.0.3",
]
